stub out URL object alloc/free
Fun JNA challenge: nng_url_parse(nng_url**, const char*)

JNA needs a little help figuring out how to handle this one because,
even though it will treat JNA Structures as pointers-to-structs when
mapping to function args, it can't handle pointer to a pointer to a
struct.

For Messages, this is easy as we just care about a Pointer...but for the
URL I wanted to expose the members of the struct.

I may bail on this design if it proves to not be worth it.