explicit unsigned types, some Size nonsense
So Unions are interesting with JNA and maybe I found a bug. Check out
the convert() methods on the UInt{16,32,64} and Size classes. It seems
that the Union reader logic doesn't do byte order correction (for
reading on LittleEndian hosts...like Intel x64). So silly!

Did some type cleanup. Better usize_t handling to be safe and correct,
though the chance of it mattering is pretty low!