hacking on aio and contexts
Got a very basic wrapper put together around Contexts using AIO. The nng
library calls are there now, as well as lower level data types, but
higher level Java access is still in flight.

For now, Contexts are available at a high level with basic send/receive
functionality with Messages, however with the caveat that it's
synchronous. Technically you could use Java's multi-threading and
concurrency on top of this, but I need to think about a better way to
mesh NNG's aio design with an established pattern in Java.