2nd prototype of event handling for Contexts
Trying to find a balancer between preserving the capabilities in the raw
Aio/Context api and a Java-centric way to access things. Unfortunately,
it's already involving a lot of proxy interfaces. Seems messy...but the
initial async "demo" seems to be working and involves minimal, yet
similar, code to the example in C.